Modularized Create An Appointment Component. Moved useEffects to parent component, prop drilled relevant state. Removed unnecessary code from child components
