+++ src/Components/NavigationBars/VisitorNavigationBar/VisitorNavigationBar.jsx
@@ -24,9 +24,9 @@ export const VisitorNavigationBar = () => {
             navigate("/", { replace: true });
           }}
         >
-          <a href="https://www.youtube.com/watch?v=dQw4w9WgXcQ">
+          {/* <a href="https://www.youtube.com/watch?v=dQw4w9WgXcQ"> */}
           Logout
-          </a>
+          {/* </a> */}
         </Link>
       </li>
     </ul>
